Remove outer content containers.
I'd initially added outer containers for the 4 sections, to then line
up the content in the inner containers. This didn't look like it was
ideal as outer containers were being re-sized. Updating to remove this,
we'll see if this works better.
Slight text edits, closer towards what I'll have in final version.