Added background colour for content boxes
I'm leaving the result as it is for now, to revisit later.
I went a bit out of scope from the design laid out in the exercise
and my vision of how it would look does not fully match how it
actually looks.
I will possibly re-visit this once I've got more experience to sharpen
it up but for the purposes of learning flexbox I've got a reasonable
understanding so I'm going to leave the site as is for now.